+++ HSPI_ZWaveParameters.csproj
@@ -87,6 +87,7 @@
     <Compile Include="Hspi\BootstrapHtmlHelper.cs" />
     <Compile Include="Hspi\DeviceConfigPage.cs" />
     <Compile Include="Hspi\HSPI.cs" />
+    <Compile Include="Hspi\OpenZWaveDB\ZWaveDevice.cs" />
     <Compile Include="Hspi\OpenZWaveDB\ZWaveDeviceManufacturer.cs" />
     <Compile Include="Hspi\OpenZWaveDB\OpenZWaveDBInformation.cs" />
     <Compile Include="Hspi\OpenZWaveDB\ZWaveDeviceParameter.cs" />
